+++ HISTORY.md
@@ -1,7 +1,13 @@
 ## HISTORY ##
 
+*v0.9.8 - 2015-08-06*
+
+* Files can now be loaded from either files
+* change default filenames to be more inline with the function
+* They are  curlyGlobal->`curlyFIXED` and curlyData->`curlyVAR`
+
 
 *v0.9.6 - 2015-07-07*
 
 * A blank files, blanks the line.
-* NOTE: A quirky in TCLSH requires the use of *continue*, instead of wiping the line.
+* NOTE: A quirky in TCLSH requires the use of *continue* (a loop), instead of wiping the line.